+++ CHANGELOG.md
@@ -21,6 +21,8 @@ If you try to use the latest Minecraft version branch, you may run into bugs and
 
 - Fixed ceiling of Desert Dungeons looking weird when it meets the wall.
 
+- Dungeons will now log error if fed an invalid identifier to a non-existent nbt file.
+
 ##### Misc:
 - Made my modifyConstant mixins to the surfacebuilders no longer crash if someone else also modifyConstant the same spot.